# server/src/services/chat/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Микросервис управления чатами и сообщениями для корпоративного мессенджера",
        "title": "Chat Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@messenger.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8084",
    "basePath": "/api/v1/chats",
    "paths": {
...
        "/chats/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех участников чата",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-members"
                ],
                "summary": "Получить список участников чата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChatMembersResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Пользователь не является участником чата",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чат не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новых участников в групповой чат или канал",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-members"
                ],
                "summary": "Добавить участников в чат",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для добавления участников",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.AddMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чат не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/chats/{id}/members/{user_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменяет роль участника чата (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-members"
                ],
                "summary": "Изменить роль участника чата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новая роль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMemberRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMemberRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидная роль",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чат или участник не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет участника из группового чата или канала (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat-members"
                ],
                "summary": "Удалить участника из чата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Чат или участник не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddMembersRequest": {
            "description": "Данные для добавления участников в чат",
            "type": "object",
            "required": [
                "role",
                "user_ids"
            ],
            "properties": {
                "role": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 1,
                    "example": 1
                },
                "user_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        4,
                        6,
                        7
                    ]
                }
            }
        },
        "models.AddMembersResponse": {
            "description": "Результат добавления участников",
            "type": "object",
            "properties": {
                "added": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        4,
                        6,
                        7
                    ]
                },
                "chat_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.ChatMemberResponse": {
            "description": "Информация об участнике чата",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "joined_at": {
                    "type": "string",
                    "example": "2024-01-01"
                },
                "login": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "patronymic": {
                    "type": "string",
                    "example": "Ivanovich"
                },
                "role": {
                    "type": "integer",
                    "example": 2
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "surname": {
                    "type": "string",
                    "example": "Ivanov"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.ChatMembersResponse": {
            "description": "Список участников чата",
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatMemberResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "models.UpdateMemberRoleRequest": {
            "description": "Новая роль участника (1=участник, 2=администратор)",
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 1,
                    "example": 2
                }
            }
        },
        "models.UpdateMemberRoleResponse": {
            "description": "Результат изменения роли",
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "type": "integer",
                    "example": 2
                },
                "user_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: Bearer {token}. Gateway проверяет токен и добавляет X-User-ID в заголовок",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}